Refactor offline TTS postprocessing
