+++ include/speech_core/tts_synthesis_options.h
@@ -29,4 +29,8 @@ std::vector<float> apply_tts_postprocess(const float* samples,
                                          int sample_rate,
                                          TtsPostProcessFlags flags);
 
+std::vector<float> apply_tts_postprocess(std::vector<float> samples,
+                                         int sample_rate,
+                                         TtsPostProcessFlags flags);
+
 }  // namespace speech_core
