+++ src/models/litert/chatterbox_c.cpp
@@ -3,7 +3,8 @@
 
 #include "speech_core/chatterbox_c.h"
 #include "speech_core/models/litert_chatterbox_tts.h"
-#include "speech_core/tts_synthesis_options.h"
+
+#include "tts_c_options.h"
 
 #include <cstdio>
 #include <exception>
@@ -11,46 +12,13 @@
 #include <string>
 
 using speech_core::LiteRTChatterboxTts;
-using speech_core::TtsSynthesisMode;
 using speech_core::TtsSynthesisOptions;
+using speech_core::convert_c_tts_synthesis_options;
 
 struct sc_chatterbox_s {
     std::unique_ptr<LiteRTChatterboxTts> tts;
     std::string last_error;
 };
-
-namespace {
-
-bool convert_options(sc_chatterbox_t h,
-                     const sc_tts_synthesis_options_t* options,
-                     TtsSynthesisOptions& out) {
-    out = TtsSynthesisOptions {};
-    if (options == nullptr) {
-        return true;
-    }
-
-    if (options->struct_size < sizeof(sc_tts_synthesis_options_t)) {
-        if (h) h->last_error = "invalid TTS synthesis options struct_size";
-        return false;
-    }
-
-    switch (options->mode) {
-    case SC_TTS_SYNTHESIS_STREAMING:
-        out.mode = TtsSynthesisMode::Streaming;
-        break;
-    case SC_TTS_SYNTHESIS_BUFFERED:
-        out.mode = TtsSynthesisMode::Buffered;
-        break;
-    default:
-        if (h) h->last_error = "unsupported TTS synthesis mode";
-        return false;
-    }
-
-    out.postprocess_flags = options->postprocess_flags;
-    return true;
-}
-
-}  // namespace
 
 sc_chatterbox_t sc_chatterbox_create(const char* bundle_dir) {
     if (!bundle_dir) return nullptr;
@@ -98,7 +66,9 @@ int sc_chatterbox_synthesize_with_options(
         return -1;
     }
     TtsSynthesisOptions cpp_options;
-    if (!convert_options(h, options, cpp_options)) {
+    std::string error;
+    if (!convert_c_tts_synthesis_options(options, cpp_options, error, "TTS")) {
+        h->last_error = error;
         return -1;
     }
     try {
