+++ src/models/voxcpm2/onnx_voxcpm2_tts.cpp
@@ -39,7 +39,7 @@ std::vector<float> apply_postprocess(
     std::vector<float> audio,
     int sample_rate,
     VoxCPM2PostProcessFlags flags) {
-    return apply_tts_postprocess(audio.data(), audio.size(), sample_rate, flags);
+    return apply_tts_postprocess(std::move(audio), sample_rate, flags);
 }
 
 }  // namespace
